+++ Products/Reportek/views/satellite_registry_management.py
@@ -4,6 +4,9 @@ from Products.Reportek.RepUtils import fix_json_from_id
 from plone.memoize.ram import global_cache
 import json
 import re
+import logging
+
+logger = logging.getLogger("Reportek")
 
 
 class SatelliteRegistryManagement(BaseAdmin):
@@ -12,16 +15,23 @@ class SatelliteRegistryManagement(BaseAdmin):
         super(SatelliteRegistryManagement, self).__call__(*args, **kwargs)
 
         api = None
+        domain = self.request.form.get('domain', 'FGAS')
         if self.request.method == "POST":
             api = self.get_api()
             if not api:
                 return None
 
         if self.request.method == "POST" and self.request.get("verify.btn"):
+            candidateId = self.request.form.get("cid")
             newId = self.request.form.get("fid")
             userId = self.request.form.get("user")
-            if newId and userId:
-                if api.verifyCandidate(newId, userId):
+            if candidateId and newId and userId:
+                if candidateId == "none":
+                    candidateId = None
+                isForMatch = api.verifyCandidate(newId, userId,
+                                                 candidateId=candidateId,
+                                                 domain=domain)
+                if isForMatch:
                     return self.request.response.redirect('{0}/{1}?done=1'.format(
                         self.context.absolute_url(), "organisation_matching"))
 
@@ -63,7 +73,7 @@ class SatelliteRegistryManagement(BaseAdmin):
             status = self.request.get('newval')
             orgid = self.request.get('orgid')
             if orgid and status:
-                if not api.updateCompanyStatus(orgid, status.upper()):
+                if not api.updateCompanyStatus(orgid, status.upper(), domain=domain):
                     return self.index(error='Unable to change company status')
                 else:
                     # We need to clear the company_details cache
@@ -79,14 +89,15 @@ class SatelliteRegistryManagement(BaseAdmin):
         api = self.get_api()
         if not api:
             return None
-        companies = api.get_registry_companies()
+        domain = self.request.form.get('domain', 'FGAS')
+        companies = api.get_registry_companies(domain=domain)
         self.request.response.setHeader('Content-Type', 'application/json')
         return json.dumps(companies, indent=2)
 
     def get_companies_api(self):
         get_params = ['id', 'vat', 'name', 'countrycode', 'OR_vat', 'OR_name']
         params = {}
-
+        domain = self.request.form.get('domain', 'FGAS')
         for param in get_params:
             if self.request.get(param):
                 if param == 'countrycode':
@@ -96,16 +107,17 @@ class SatelliteRegistryManagement(BaseAdmin):
 
         api = self.get_api()
         self.request.response.setHeader('Content-Type', 'application/json')
-        return json.dumps(api.existsCompany(params), indent=2)
+        return json.dumps(api.existsCompany(params, domain=domain), indent=2)
 
     def get_company_json(self):
         self.request.response.setHeader('Content-Type', 'application/json')
         api = self.get_api()
-
+        domain = self.request.form.get('domain', 'FGAS')
         details = {}
         if self.request.get('id'):
             companyId = self.request.get('id')
-            details = fix_json_from_id(api.get_company_details(companyId))
+            details = fix_json_from_id(api.get_company_details(companyId,
+                                                               domain=domain))
 
         return json.dumps(details, indent=2)
 
@@ -118,7 +130,8 @@ class SatelliteRegistryManagement(BaseAdmin):
             return None
 
         companyId = self.request.get('id')
-        data = api.getCompanyDetailsById(companyId)
+        domain = self.request.form.get('domain', 'FGAS')
+        data = api.getCompanyDetailsById(companyId, domain=domain)
         if not data:
             return None
 
@@ -135,8 +148,8 @@ class SatelliteRegistryManagement(BaseAdmin):
         api = self.get_api()
         if not api:
             return None
-
-        candidates = api.getCandidates()
+        domain = self.request.form.get('domain', 'FGAS')
+        candidates = api.getCandidates(domain=domain)
         self.request.response.setHeader('Content-Type', 'application/json')
         return json.dumps(candidates, indent=2)
 
@@ -144,13 +157,19 @@ class SatelliteRegistryManagement(BaseAdmin):
         api = self.get_api()
         if not api:
             return None
-        return api.getMatchingLog()
+        domain = self.request.form.get('domain', 'FGAS')
+        m_logs = api.getMatchingLog(domain=domain)
+        self.request.response.setHeader('Content-Type', 'application/json')
+        return json.dumps(m_logs, indent=2)
 
     def get_datasync_log(self):
         api = self.get_api()
         if not api:
             return None
-        return api.getDataSyncLog()
+        domain = self.request.form.get('domain', 'FGAS')
+        sync_logs = api.getDataSyncLog(domain=domain)
+        self.request.response.setHeader('Content-Type', 'application/json')
+        return json.dumps(sync_logs, indent=2)
 
     def unverify(self):
         details = {}
@@ -158,7 +177,8 @@ class SatelliteRegistryManagement(BaseAdmin):
             api = self.get_api()
             if api:
                 companyId = self.request.get('id')
-                details = api.unverifyCompany(companyId, self.request.AUTHENTICATED_USER.getUserName())
+                domain = self.request.form.get('domain', 'FGAS')
+                details = api.unverifyCompany(companyId, self.request.AUTHENTICATED_USER.getUserName(), domain=domain)
         return json.dumps(details, indent=2)
 
     def get_companies_excel(self):
@@ -172,15 +192,17 @@ class SatelliteRegistryManagement(BaseAdmin):
         api = self.get_api()
         if not api:
             return None
-        response = api.getCompaniesExcelExport()
+        domain = self.request.form.get('domain', 'FGAS')
+        response = api.getCompaniesExcelExport(domain=domain)
         return response.content
 
     def get_companies_json(self):
         api = self.get_api()
         if not api:
             return None
-
-        companies_original = api.get_registry_companies(detailed=True)
+        domain = self.request.form.get('domain', 'FGAS')
+        companies_original = api.get_registry_companies(detailed=True,
+                                                        domain=domain)
         if companies_original is None:
             companies_original = []
 
@@ -228,6 +250,15 @@ class SatelliteRegistryManagement(BaseAdmin):
             old_collection_id = None
         return api.lockedCompany(str(company_id), old_collection_id,
                                  country_code, domain)
+
+    def is_company_locked(self):
+        company_id = self.request.form.get('company_id')
+        old_collection_id = self.request.form.get('old_collection_id')
+        country_code = self.request.form.get('country_code')
+        domain = self.request.form.get('domain', 'FGAS')
+        self.request.response.setHeader('Content-Type', 'application/json')
+        return json.dumps(self.lockedCompany(company_id, old_collection_id,
+                                             country_code, domain))
 
     def lockDownCompany(self, company_id, old_collection_id, country_code, domain, user, came_from):
         api = self.get_api()